+++ packages/codemod/src/transforms/internal/core/alert/__testfixtures__/alert-with-svg.output.tsx
@@ -1,8 +1,7 @@
-// @ts-nocheck
 import { Callout } from '@vapor-ui/core';
 
 export const Component = () => (
-    <Callout.Root color="warning">
+    <Callout.Root colorPalette="warning">
         <Callout.Icon>
             <svg width="20" height="20" viewBox="0 0 20 20">
                 <path d="M10 0L20 20H0L10 0Z" />