+++ packages/codemod/src/transforms/internal/core/avatar/__testfixtures__/avatar-with-spread.input.tsx
@@ -1,0 +1,6 @@
+import { Avatar } from '@goorm-dev/vapor-core';
+
+function Component() {
+    const props = { size: 'md' as const };
+    return <Avatar label="User" {...props} />;
+}