+++ packages/codemod/src/transforms/internal/core/radio-group/__testfixtures__/radio-group-merge-imports.output.tsx
@@ -1,0 +1,12 @@
+import { Button, Radio, RadioGroup } from '@vapor-ui/core';
+
+export function Example() {
+    return (
+        <>
+            <RadioGroup.Root defaultValue="1">
+                <Radio.Root value="1" />
+            </RadioGroup.Root>
+            <Button>Submit</Button>
+        </>
+    );
+}