+++ packages/codemod/src/transforms/internal/core/text-input/__testfixtures__/text-input-ref.output.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+import { Field, TextInput } from '@vapor-ui/core';
+
+export function RefExample() {
+    const inputRef = React.useRef<HTMLInputElement>(null);
+
+    return (
+        <Field.Root>
+            <Field.Label>
+                Name
+                <TextInput size="md" ref={inputRef} placeholder="Enter name" />
+            </Field.Label>
+        </Field.Root>
+    );
+}