+++ packages/codemod/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @vapor-ui/codemod
 
+## 1.0.0-beta.7
+
+### Major Changes
+
+- 2c8f651: Add Tabs, Text, Textinput transform
+- 441e481: Add Dropdown, IconButton, Nav transformer
+- 441e481: add avatar, breadcrumb, card transform
+- 441e481: Add Popover, RadioGroup, Switch Transform
+
+### Minor Changes
+
+- 441e481: Add codemod for icon packages name import
+- 441e481: add codemod cli packages
+
+### Patch Changes
+
+- 15a2f7c: add cli options
+- 88da973: add button, badge, alert component color props to colorPalette
+
 ## 1.0.0-beta.6
 
 ### Minor Changes